feat: nfts display on user profile with number in user wallet
